requirements.txt
streamlit
transformers
torch
pillow
numpy
requests
